refactor: cardHideOptionSelect feature로 분리
