--- src/features/drilling/components/cardHideOptionSelect/index.test.tsx
@@ -1,3 +1,0 @@
-import { test } from 'vitest';
-
-test('CardHideOptionSelect listbox renders card hide options', () => {});
