+++ src/features/drilling/types/cardHideOption.types.ts
@@ -1,6 +1,6 @@
 import { ApiResultType } from '@/shared/types/apiResult.type';
 import { ArrayElement } from '@/shared/types/arrayElement.type';
-import { getCardHideOptions } from '@features/drilling/api/getCardHideOptions';
+import { getCardHideOptions } from '@features/cardHideOptionSelect/api/getCardHideOptions';
 
 export type CardHideOptionTypes = ApiResultType<typeof getCardHideOptions>;
 export type CardHideOption = ArrayElement<CardHideOptionTypes>;
